# api/openapi.yaml
openapi: 3.1.0
info:
  title: Arca
  version: "1"
  description: "Durable storage for people, agents, and sandboxes: files with versions, trash and stars, uploads in parts, shares with a permission ladder and public links, workspaces with a writer lease, and an event log."
paths:
...
  /v1/workspaces:
    get:
      operationId: getV1Workspaces
      summary: List workspaces
      description: List the workspaces of a space. Asks the authorizer for workspace.list before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postV1Workspaces
      summary: Create workspace
      description: Create a workspace in a space. Asks the authorizer for workspace.create before it acts.
      security:
      - bearer: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/deleted:
    get:
      operationId: getV1WorkspacesDeleted
      summary: List deleted workspaces
      description: List the soft deleted workspaces of a space that are still restorable. Asks the authorizer for workspace.list before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}:
    delete:
      operationId: deleteV1WorkspacesById
      summary: Delete workspace
      description: Soft delete a workspace, leaving it restorable. Asks the authorizer for workspace.delete before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: getV1WorkspacesById
      summary: Get workspace
      description: Read one workspace with its lease and the counters of its subtree. Asks the authorizer for workspace.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: patchV1WorkspacesById
      summary: Rename workspace
      description: Rename a workspace, which moves its rows and no bucket key. Asks the authorizer for workspace.write before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/attach:
    post:
      operationId: postV1WorkspacesByIdAttach
      summary: Attach workspace
      description: Open an attachment, taking the writer lease for a rw mount. Asks the authorizer for workspace.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/attach/{aid}:
    delete:
      operationId: deleteV1WorkspacesByIdAttachByAid
      summary: Release attachment
      description: Release an attachment and the lease it held. Asks the authorizer for workspace.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: aid
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/attach/{aid}/renew:
    post:
      operationId: postV1WorkspacesByIdAttachByAidRenew
      summary: Renew attachment
      description: Push the attachment's deadline forward. Asks the authorizer for workspace.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: aid
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/materialize:
    get:
      operationId: getV1WorkspacesByIdMaterialize
      summary: Get workspace manifest
      description: The attachment's pinned manifest with a presigned read per file. Asks the authorizer for workspace.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/restore:
    post:
      operationId: postV1WorkspacesByIdRestore
      summary: Restore workspace
      description: Undo a soft delete. Asks the authorizer for workspace.restore before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/workspaces/{id}/sync:
    post:
      operationId: postV1WorkspacesByIdSync
      summary: Sync workspace
      description: Declare the post-state manifest of the subtree; the server reconciles. Asks the authorizer for workspace.sync before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A token from an issuer this installation lists, carrying the audience it verifies.
  schemas:
    Error:
      type: object
      description: The one error shape of this API. A caller branches on error.code and shows error.message.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: The stable identifier of the row of the error table.
            details:
              type: object
              description: "What varies: always request_id, the developer sentence in detail, and the field paths at fault in fields."
              properties:
                detail:
                  type: string
                fields:
                  type: array
                  items:
                    type: string
                request_id:
                  type: string
            message:
              type: string
              description: The one sentence of that row, fixed and shown to a person.
          required:
          - code
          - message
      required:
      - error
...
  responses:
...
    forbidden:
      description: "403 forbidden: You do not have permission to do this."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    internal:
      description: "500 internal: Something went wrong on this server."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    not_found:
      description: "404 not_found: There is no such object."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    rate_limited:
      description: "429 rate_limited: Too many requests; wait and retry."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    unauthenticated:
      description: "401 unauthenticated: Sign in and send a valid token."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...